@wip
so much trash! Trying to diagnose what is going on in checkout_spec.
Here's what I know:

1. The specs block starting at line 256 passes in isolation but fails in
`bundle exec rspec --seed 41665`. This tells me that an earlier test is
setting something that breaks this one.
2. The failure for this test is likely due to switching line 14 from
`build_stubbed_list` to `create_list` which causes the installments to
be processed.
3. Within Solidus, in `app/models/spree/payment_method/store_credit.rb`
there is a `purchase` method that searches for an eligible store credit
via its events. The logic there is a bit opaque, but it seems to be
looking for a credit whose only event is one of type "eligible". The
credit that gets created in the line 265 spec block has a second event,
of type "allocation". I haven't yet figured out why this exists or why
it seems to disqualify the store credit from use.